Single xaudio2 source file, fix vector calls in asio.cpp
